refactor: codec and rules organisation
