+++ src/codecs/domain/fares.rules.ts
@@ -1,8 +1,8 @@
+/* eslint-disable @typescript-eslint/typedef */
 import { intersection as ioIntersection } from 'io-ts';
-import { driveRulesCodec, durationDistanceRulesCodec, passengerRulesCodec } from './traits.codecs';
 import { fareToEditCodec, returnDriveCodec, toEditCodec, toScheduleCodec } from './fares.codecs';
+import { driveRulesCodec, durationDistanceRulesCodec, passengerRulesCodec } from './traits.rules';
 
-// eslint-disable-next-line @typescript-eslint/typedef
 export const toScheduleRulesCodec = ioIntersection([
   toScheduleCodec,
   driveRulesCodec,
@@ -10,13 +10,10 @@ export const toScheduleRulesCodec = ioIntersection([
   durationDistanceRulesCodec
 ]);
 
-// eslint-disable-next-line @typescript-eslint/typedef
 export const returnDriveRulesCodec = ioIntersection([returnDriveCodec, driveRulesCodec, durationDistanceRulesCodec]);
 
-// eslint-disable-next-line @typescript-eslint/typedef
 export const toEditRulesCodec = ioIntersection([toEditCodec, driveRulesCodec, passengerRulesCodec, durationDistanceRulesCodec]);
 
-// eslint-disable-next-line @typescript-eslint/typedef
 export const fareToEditRulesCodec = ioIntersection([
   fareToEditCodec,
   driveRulesCodec,
