+++ src/codecs/domain/index.ts
@@ -1,4 +1,6 @@
 export * from './traits.codecs';
+export * from './traits.rules';
 export * from './regular.codecs';
+export * from './regular.rules';
 export * from './fares.codecs';
 export * from './fares.rules';
