+++ src/codecs/domain/regular.codecs.ts
@@ -1,6 +1,6 @@
 import { intersection as ioIntersection, string as ioString, type as ioType, Type, array as ioArray } from 'io-ts';
 import { Entity, Regular } from '../../definitions';
-import { entityCodec, passengerRulesCodec } from './traits.codecs';
+import { entityCodec } from './traits.codecs';
 
 export const regularPassengerCodec: Type<Regular> = ioType({
   firstname: ioString,
@@ -8,10 +8,6 @@ export const regularPassengerCodec: Type<Regular> = ioType({
   phone: ioString
 });
 
-// eslint-disable-next-line @typescript-eslint/typedef
-export const regularPassengerRulesCodec = ioIntersection([regularPassengerCodec, passengerRulesCodec]);
-
 export const regularPassengerEntityCodec: Type<Entity & Regular> = ioIntersection([entityCodec, regularPassengerCodec]);
 
-export const regularsCodec: Type<(Entity & Regular)[]> = ioArray(regularPassengerEntityCodec)
-
+export const regularsCodec: Type<(Entity & Regular)[]> = ioArray(regularPassengerEntityCodec);
